useModalController 공용 hook 분리 -> useFetchAndUpdate hook 생성
